Add CRLF character to end of multipart body.
Per RFC 1521 this character is required at the end of multipart body.
Per RFC 2046 there is an optional CRLF character at the end
of the multipart-body to delimit between the body and optional
epilogue.

Some server implementations break if this character is
not added, so adding it here as it complies with both RFCs.